--- pkg/gomapper/tests/map_test.go
@@ -1,1 +1,0 @@
-package tests
