+++ pkg/gomapper/tests/maptype/maptype_test.go
@@ -1,0 +1,1 @@
+package maptype
