+++ pkg/gomapper/tests/slicetype/slicetype_test.go
@@ -1,4 +1,4 @@
-package slice
+package slicetype
 
 import (
 	"testing"
